+++ src/main/java/lotto/domain/LottoStore.java
@@ -1,6 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import calculator.Operation;
+import lotto.domain.Lotto;
+import lotto.domain.LottoStorage;
 
 import java.util.List;
