+++ src/test/java/lotto/LottoStorageTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoStorage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
